Implement tests for GetAllAdmins
